pyproject.toml
[project]
name = "l2c-ctfs-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "asyncio>=3.4.3",
    "azure-functions>=1.22.1",
    "mcp>=1.6.0",
    "python-dotenv>=1.1.0",
]
